requirements.txt
bleach==2.1.3
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.3
django-allauth==0.35.0
django-bleach==0.3.0
django-bootstrap-form==3.4
django-crispy-forms==1.7.2
django-debug-toolbar==1.9.1
django-markdown-deux==1.0.5
django-rosetta==0.8.1
docutils==0.14
gunicorn==19.7.1
html5lib==1.0.1
idna==2.6
markdown2==2.3.5
microsofttranslator==0.8
oauthlib==2.0.7
Pillow==5.0.0
pkg-resources==0.0.0
polib==1.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
python-dateutil==2.7.2
python-memcached==1.59
python3-openid==3.1.0
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
sqlparse==0.2.4
urllib3==1.22
webencodings==0.5.1
